# Conflicts:
#	pos-people/src/main/resources/db/migration/R__seed_timekeeping_approval_data.sql

+++ pos-shop-manager/src/main/java/com/positivity/shopmanager/internal/entity/Mechanic.java
@@ -55,8 +55,10 @@ public class Mechanic {
     @Column(name = "termination_date")
     private LocalDate terminationDate;
 
+    // HR-supplied monotonic sync marker (people.events.v1 aggregateVersion is epoch millis),
+    // not a JPA optimistic lock.
     @Column(name = "version")
-    private int version;
+    private long version;
 
     @Column(name = "last_synced_at")
     private Instant lastSyncedAt;
